Getting closer with api refactor
