+++ StompboxAPI/UnmanagedAudioPlugin.cs
@@ -2,6 +2,7 @@
 using System.Collections.Generic;
 using System.Collections.ObjectModel;
 using System.Runtime.InteropServices;
+using Stompbox;
 
 namespace StompboxAPI
 {
