+++ StompboxAPIBase/StompboxAPIBase.projitems
@@ -10,6 +10,7 @@
   </PropertyGroup>
   <ItemGroup>
     <Compile Include="$(MSBuildThisFileDirectory)AudioPlugin.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)PluginFactory.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)StompboxClient.cs" />
   </ItemGroup>
 </Project>
