+++ StompboxAndroid/StompboxAndroid.csproj
@@ -20,6 +20,9 @@
     <PackageReference Include="MonoGame.Content.Builder.Task" Version="3.8.2.1105" />
     <PackageReference Include="MonoGame.Framework.Android" Version="3.8.2.1105" />
   </ItemGroup>
+  <ItemGroup>
+    <ProjectReference Include="..\StompboxRemoteClient\StompboxRemoteClient.csproj" />
+  </ItemGroup>
   <Target Name="RestoreDotnetTools" BeforeTargets="Restore">
     <Message Text="Restoring dotnet tools" Importance="High" />
     <Exec Command="dotnet tool restore" />
