+++ StompboxPlugin/StompboxPlugin.csproj
@@ -10,6 +10,7 @@
 		<PlatformTarget>AnyCPU</PlatformTarget>
 	</PropertyGroup>
 	<ItemGroup>
+		<ProjectReference Include="..\StompboxAPI\StompboxAPI.csproj" />
 		<ProjectReference Include="..\UnmanagedBridge\UnmanagedBridge.vcxproj" />
 	</ItemGroup>
 	<ItemGroup>
